some more tests

+++ db/migrations/20210503111934_create_goals_table.js
@@ -7,7 +7,7 @@ exports.up = function(knex, Promise) {
     table.integer('carbs');
     table.timestamp('created_at').defaultTo(knex.fn.now());
     table.timestamp('updated_at').defaultTo(knex.fn.now());
-    table.bigInteger('userId').references('id').inTable('users');
+    table.integer('userId').references('id').inTable('users');
   });
 };
